rv32-emu : support booting Linux 5.4
This commit can support booting Linux 5.4 but
the init process will fail because the emulator
need to support F/D-ext.
(or Linux and busybox built without F/D instruction)

The emulator still have a lot of bugs and
also need to refactor to let code clearly and
get better performance.
    - FIXME: Timer bug causes rv32-xv6 booting fail.

Signed-off-by: Yen <qqgnoe466263@gmail.com>